% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo.R
\name{ta_elo_win_prob}
\alias{ta_elo_win_prob}
\title{Expected win probability from two Elo ratings}
\usage{
ta_elo_win_prob(elo_a, elo_b, scale = 400)
}
\arguments{
\item{elo_a, elo_b}{Ratings.}

\item{scale}{Elo scale (400).}
}
\value{
The probability that \code{a} beats \code{b}.
}
\description{
Expected win probability from two Elo ratings
}
